Fixed all paths after restructure
