+++ src/App/App.jsx
@@ -4,18 +4,18 @@ import { Provider } from 'react-redux';
 import configureStore from './configureStore';
 import 'bootstrap';
 import History from './history';
-import NavbarConnected from './components/navbar';
-import FooterConnected from './components/footer';
-import LandingPageConnected from './components/landingPage';
-import DashboardConnected from './components/dashboard';
-import UploadScreenConnected from './components/uploadScreen';
-import LinkoutPageConnected from './components/linkoutPage';
-import AnalysisHomePageConnected from './components/analysisHomePage';
-import DownloadPageConnected from './components/downloadPage';
-import MethodsConnected from './components/methods';
-import Team from './components/team';
-import Contact from './components/contact';
-import CallbackConnected from './components/callback';
+import NavbarConnected from '../Navbar/navbar';
+import FooterConnected from '../Footer/footer';
+import LandingPageConnected from '../LandingPage/landingPage';
+import DashboardConnected from '../Dashboard/dashboard';
+import UploadScreenConnected from '../UploadPage/uploadScreen';
+import LinkoutPageConnected from '../LinkoutPage/linkoutPage';
+import AnalysisHomePageConnected from '../AnalysisPage/analysisHomePage';
+import DownloadPageConnected from '../DownloadPage/downloadPage';
+import MethodsConnected from '../MethodsPage/methods';
+import Team from '../TeamPage/team';
+import Contact from '../ContactPage/contact';
+import CallbackConnected from '../Auth/callback';
 
 const store = configureStore();
