+++ src/Auth/__test__/authLoading.stories.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { AuthLoading } from '../components/authLoading';
+import { AuthLoading } from '../authLoading';
 
 const authActions = {
   authSuccess: action('Authorized!'),
