+++ src/DownloadPage/__test__/downloadRow.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Adapter from 'enzyme-adapter-react-16';
 import Enzyme, { shallow } from 'enzyme';
-import DownloadRow from '../../components/downloadRow';
+import DownloadRow from '../downloadRow';
 
 Enzyme.configure({ adapter: new Adapter() });
