+++ src/Footer/footer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
-import actions from '../actions';
+import actions from '../Auth/authActions';
 
 /**
  * Method to render global footer
