+++ src/LandingPage/__test__/landingPage.stories.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { LandingPage } from '../components/landingPage';
-import { Navbar } from '../components/navbar';
-import { Footer } from '../components/footer';
+import { LandingPage } from '../landingPage';
+import { Navbar } from '../../Navbar/navbar';
+import { Footer } from '../../Footer/footer';
 
 const footerActions = {
   login: action('logging in'),
