+++ src/LinkoutPage/__test__/linkoutPage.stories.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { LinkoutPage } from '../components/linkoutPage';
-import { Navbar } from '../components/navbar';
-import { Footer } from '../components/footer';
+import { LinkoutPage } from '../linkoutPage';
+import { Navbar } from '../../Navbar/navbar';
+import { Footer } from '../../Footer/footer';
 
 const footerActions = {
   login: action('logging in'),
