+++ src/Navbar/__test__/navbar.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { Navbar } from '../components/navbar';
+import { Navbar } from '../navbar';
 
 storiesOf('Navbar', module)
   .add('default', () => <Navbar />)
