+++ src/Navbar/__test__/navbar.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Adapter from 'enzyme-adapter-react-16';
 import Enzyme, { shallow } from 'enzyme';
-import { Navbar } from '../../components/navbar';
+import { Navbar } from '../navbar';
 
 Enzyme.configure({ adapter: new Adapter() });
