+++ src/Widgets/__test__/dashboardWidgets.stories.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import PreviousUploadsGraph from '../components/previousUploadsGraph';
-import { PreviousUploadsTable } from '../components/previousUploadsTable';
-import AllUploadsDoughnut from '../components/allUploadsDoughnut';
-import AllUploadStats from '../components/allUploadStats';
+import PreviousUploadsGraph from '../previousUploadsGraph';
+import { PreviousUploadsTable } from '../previousUploadsTable';
+import AllUploadsDoughnut from '../allUploadsDoughnut';
+import AllUploadStats from '../allUploadStats';
 
-const previousUploads = require('../testData/testPreviousUploads');
-const allUploads = require('../testData/testAllUploads');
+const previousUploads = require('../../testData/testPreviousUploads');
+const allUploads = require('../../testData/testAllUploads');
 
 const expandedUploads = previousUploads.map(upload => ({
   ...upload,
